PEP8 cleaning, create package for tilesets
Other changes include adapting tiles.py to use updated seek_class event, refactoring MapTile.available_actions() so it is easier to read
